fix(web): nested layout navigation issue
